Refactor body adapter itemonclick into BodyAdapter.java
