+++ app/src/main/java/app/insti/adapter/BodyAdapter.java
@@ -1,6 +1,9 @@
 package app.insti.adapter;
 
 import android.content.Context;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentTransaction;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,25 +11,26 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.google.gson.Gson;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
 import app.insti.Constants;
-import app.insti.interfaces.ItemClickListener;
 import app.insti.R;
 import app.insti.api.model.Body;
+import app.insti.fragment.BodyFragment;
 
 
 public class BodyAdapter extends RecyclerView.Adapter<BodyAdapter.ViewHolder> {
 
     private List<Body> bodyList;
-    private ItemClickListener itemClickListener;
     private Context context;
+    private Fragment fragment;
 
-    public BodyAdapter(List<Body> bodyList, ItemClickListener itemClickListener) {
+    public BodyAdapter(List<Body> bodyList, Fragment mFragment) {
         this.bodyList = bodyList;
-        this.itemClickListener = itemClickListener;
+        fragment = mFragment;
     }
 
     @Override
@@ -39,11 +43,26 @@ public class BodyAdapter extends RecyclerView.Adapter<BodyAdapter.ViewHolder> {
         v.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                itemClickListener.onItemClick(view, postViewHolder.getAdapterPosition());
+                openBody(bodyList.get(postViewHolder.getAdapterPosition()));
             }
         });
 
         return postViewHolder;
+    }
+
+    /**
+     * Open body fragment for a body
+     */
+    private void openBody(Body body) {
+        Bundle bundle = new Bundle();
+        bundle.putString(Constants.BODY_JSON, new Gson().toJson(body));
+        BodyFragment bodyFragment = new BodyFragment();
+        bodyFragment.setArguments(bundle);
+        FragmentTransaction ft = fragment.getActivity().getSupportFragmentManager().beginTransaction();
+        ft.setCustomAnimations(R.anim.slide_in_left, R.anim.slide_out_left, R.anim.slide_in_right, R.anim.slide_out_right);
+        ft.replace(R.id.framelayout_for_fragment, bodyFragment, bodyFragment.getTag());
+        ft.addToBackStack(bodyFragment.getTag());
+        ft.commit();
     }
 
     @Override
